--- a/examples/text_to_image_sdxl.py
+++ b/examples/text_to_image_sdxl.py
@@ -46,83 +46,87 @@ cmd_args = parser.parse_args()
 if cmd_args.compile and cmd_args.compile_with_dynamo:
     parser.error("--compile and --compile_with_dynamo cannot be used together.")
 
-# Normal SDXL pipeline init.
-seed = torch.Generator("cuda").manual_seed(cmd_args.seed)
-output_type = "pil"
-# SDXL base: StableDiffusionXLPipeline
-base = DiffusionPipeline.from_pretrained(
-    cmd_args.base,
-    torch_dtype=torch.float16,
-    variant=cmd_args.variant,
-    use_safetensors=True,
-)
-base.to("cuda")
-if cmd_args.with_refiner:
-    # SDXL refiner: StableDiffusionXLImg2ImgPipeline
-    refiner = DiffusionPipeline.from_pretrained(
-        cmd_args.refiner,
-        text_encoder_2=base.text_encoder_2,
-        vae=base.vae,
+def test_sdxl(device):
+    # Normal SDXL pipeline init.
+    seed = torch.Generator("cuda").manual_seed(cmd_args.seed)
+    output_type = "pil"
+    # SDXL base: StableDiffusionXLPipeline
+    base = DiffusionPipeline.from_pretrained(
+        cmd_args.base,
         torch_dtype=torch.float16,
-        use_safetensors=True,
         variant=cmd_args.variant,
+        use_safetensors=True,
     )
-    refiner.to("cuda")
-
-# Compile unet with oneflow
-if cmd_args.compile:
-    print("unet is compiled to oneflow.")
-    base.unet = oneflow_compile(
-        base.unet, options={"size": cmd_args.num_dynamic_input_size}
-    )
+    base.to(device)
     if cmd_args.with_refiner:
-        refiner.unet = oneflow_compile(
-            refiner.unet, options={"size": cmd_args.num_dynamic_input_size}
+        # SDXL refiner: StableDiffusionXLImg2ImgPipeline
+        refiner = DiffusionPipeline.from_pretrained(
+            cmd_args.refiner,
+            text_encoder_2=base.text_encoder_2,
+            vae=base.vae,
+            torch_dtype=torch.float16,
+            use_safetensors=True,
+            variant=cmd_args.variant,
         )
-        output_type = "latent"
-
-# Load compiled unet with oneflow
-if cmd_args.compile and cmd_args.load:
-    print("loading graphs...")
-    base.unet.warmup_with_load("base_" + cmd_args.file)
-    if cmd_args.with_refiner:
-        refiner.unet.warmup_with_load("refiner_" + cmd_args.file)
-
-# Compile unet with torch.compile to oneflow. Note this is at alpha stage(experimental) and may be changed later.
-if cmd_args.compile_with_dynamo:
-    logger.info("unet is compiled to oneflow with torch.compile.")
-    from onediff.infer_compiler import oneflow_backend
-
-    base.unet = torch.compile(
-        base.unet, fullgraph=True, mode="reduce-overhead", backend=oneflow_backend
-    )
-
-# Normal SDXL run
-# sizes = [1024, 896, 768]
-sizes = [1024]
-for h in sizes:
-    for w in sizes:
-        for i in range(3):
-            image = base(
-                prompt=cmd_args.prompt,
-                height=h,
-                width=w,
-                generator=seed,
-                num_inference_steps=cmd_args.n_steps,
-                output_type=output_type,
-            ).images
-            if cmd_args.with_refiner:
-                image = refiner(
+        refiner.to("cuda")
+    
+    # Compile unet with oneflow
+    if cmd_args.compile:
+        print("unet is compiled to oneflow.")
+        base.unet = oneflow_compile(
+            base.unet, options={"size": cmd_args.num_dynamic_input_size}
+        )
+        if cmd_args.with_refiner:
+            refiner.unet = oneflow_compile(
+                refiner.unet, options={"size": cmd_args.num_dynamic_input_size}
+            )
+            output_type = "latent"
+    
+    # Load compiled unet with oneflow
+    if cmd_args.compile and cmd_args.load:
+        print("loading graphs...")
+        base.unet.warmup_with_load("base_" + cmd_args.file + device)
+        if cmd_args.with_refiner:
+            refiner.unet.warmup_with_load("refiner_" + cmd_args.file + device)
+    
+    # Compile unet with torch.compile to oneflow. Note this is at alpha stage(experimental) and may be changed later.
+    if cmd_args.compile_with_dynamo:
+        logger.info("unet is compiled to oneflow with torch.compile.")
+        from onediff.infer_compiler import oneflow_backend
+    
+        base.unet = torch.compile(
+            base.unet, fullgraph=True, mode="reduce-overhead", backend=oneflow_backend
+        )
+    
+    # Normal SDXL run
+    # sizes = [1024, 896, 768]
+    sizes = [1024]
+    for h in sizes:
+        for w in sizes:
+            for i in range(3):
+                image = base(
                     prompt=cmd_args.prompt,
+                    height=h,
+                    width=w,
                     generator=seed,
                     num_inference_steps=cmd_args.n_steps,
-                    image=image,
+                    output_type=output_type,
                 ).images
-            image[0].save(f"h{h}-w{w}-i{i}-{cmd_args.saved_image}")
+                if cmd_args.with_refiner:
+                    image = refiner(
+                        prompt=cmd_args.prompt,
+                        generator=seed,
+                        num_inference_steps=cmd_args.n_steps,
+                        image=image,
+                    ).images
+                image[0].save(f"h{h}-w{w}-i{i}-{cmd_args.saved_image}")
+    
+    # Save compiled unet with oneflow
+    if cmd_args.compile and cmd_args.save:
+        print("saving graphs...")
+        base.unet.save_graph("base_" + cmd_args.file + device)
+        if cmd_args.with_refiner:
+            refiner.unet.save_graph("refiner_" + cmd_args.file + device)
 
-# Save compiled unet with oneflow
-if cmd_args.compile and cmd_args.save:
-    print("saving graphs...")
-    base.unet.save_graph("base_" + cmd_args.file)
-    if cmd_args.with_refiner:
-        refiner.unet.save_graph("refiner_" + cmd_args.file)
+test_sdxl("cuda")
+test_sdxl("cuda:1")