--- a/src/onediff/infer_compiler/with_oneflow_compile.py
+++ b/src/onediff/infer_compiler/with_oneflow_compile.py
@@ -32,25 +32,8 @@ def get_unet_graph(size=9):
             # os.environ["ONEFLOW_MLIR_FUSE_KERNEL_LAUNCH"] = "1"
             # os.environ["ONEFLOW_KERNEL_ENABLE_CUDA_GRAPH"] = "1"
 
-        def build(
-            self,
-            latent_model_input,
-            t,
-            encoder_hidden_states,
-            cross_attention_kwargs=None,
-            added_cond_kwargs=None,
-            return_dict=False,
-        ):
-            encoder_hidden_states = flow._C.amp_white_identity(encoder_hidden_states)
-            pred = self.unet(
-                latent_model_input,
-                t,
-                encoder_hidden_states=encoder_hidden_states,
-                cross_attention_kwargs=cross_attention_kwargs,
-                added_cond_kwargs=added_cond_kwargs,
-                return_dict=return_dict,
-            )
-            return pred
+        def build(self, *args, **kwargs):
+            return self.unet(*args, **kwargs)
 
         def warmup_with_load(self, file_path):
             state_dict = flow.load(file_path)